Indoor_Pathloss_Sajid_Final_WN/requirements.txt
Brotli==1.1.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
imageio==2.36.0
inflate64==1.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.21.0
pyparsing==3.2.0
pyppmd==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzstd==0.16.2
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sympy==1.13.1
texttable==1.7.0
tifffile==2024.9.20
torch==2.5.0
torchsummary==1.5.1
torchvision==0.20.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
